Move StringSetLintOption to util
