+++ slack-lint-checks/src/main/java/slack/lint/util/StringSetLintOption.kt
@@ -1,6 +1,6 @@
 // Copyright (C) 2023 Slack Technologies, LLC
 // SPDX-License-Identifier: Apache-2.0
-package slack.lint.compose
+package slack.lint.util
 
 import com.android.tools.lint.client.api.Configuration
 import com.android.tools.lint.detector.api.StringOption
